% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{stat_mode}
\alias{stat_mode}
\title{Statistical mode}
\usage{
stat_mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector (numeric, character, or factor).}

\item{na.rm}{Logical. Should missing values be removed before computing
the mode? Defaults to TRUE.}
}
\value{
The most frequent value of `x`.
If `x` is a factor, the result is returned as a factor with the same levels.
If there are ties, the first occurring mode is returned.
If all values are missing, returns `NA`.
}
\description{
Computes the most frequent (modal) value of a vector.
}
\examples{
stat_mode(c(1, 2, 2, 3, NA))
stat_mode(c("a", "b", "a", "c", "c"))
stat_mode(factor(c("low", "high", "low", NA)))

}
