% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mutate_code.R
\name{write_outcome_missingness_dummies_code}
\alias{write_outcome_missingness_dummies_code}
\title{Write outcome missingness code}
\usage{
write_outcome_missingness_dummies_code(data, ...)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{Unquoted variable names to generate missingness dummy variables for}
}
\value{
Invisibly returns the generated code as a single string.
}
\description{
Generates tidyverse-style `mutate()` code generate missingness
dummy variables (with suffix `_missing`) is created for selected variables.
}
\details{
This function prints mutate code to the console that you can copy–paste
into your cleaning script. It does not create the variables itself.
}
\examples{
# Example data with missingness
dat <- data.frame(
   Y_attitude = rep(c(1, 2, 3, 4, 5, NA), c(10, 20, 30, 40, 50, 50)),
   Y_behavior = rep(c(0, 1, NA), c(100, 50, 50))
)

# Generate imputation code
write_covariate_imputation_code(dat, X_factor_variable, X_numeric_variable)

}
